+++ features/user/chat_common/src/main/java/com/tokopedia/chat_common/data/ProductAttachmentViewModel.kt
@@ -35,9 +35,9 @@ open class ProductAttachmentViewModel : SendableViewModel,
         private set
     var canShowFooter: Boolean = false
         private set
-    var blastId: Long = 0L
+    var blastId: Long = 0
         private set
-    var priceInt: Int = 0
+    var priceInt: Long = 0
         private set
     var category: String = ""
         private set
@@ -45,10 +45,10 @@ open class ProductAttachmentViewModel : SendableViewModel,
         private set
     var priceBefore: String = ""
         private set
-    var shopId: Int = 0
+    var shopId: Long = 0
     var freeShipping: FreeShipping = FreeShipping()
     var playStoreData: PlayStoreData = PlayStoreData()
-    var categoryId: Int = 0
+    var categoryId: Long = 0
     var minOrder: Int = 1
     var variants: List<AttachmentVariant> = emptyList()
     var colorVariantId: String = ""
@@ -74,7 +74,7 @@ open class ProductAttachmentViewModel : SendableViewModel,
             return priceBefore.isNotEmpty() && dropPercentage.isNotEmpty()
         }
     val stringBlastId: String get() = blastId.toString()
-    var campaignId: Int = 0
+    var campaignId: Long = 0
 
     override fun updateData(attribute: Any?) {
         if (attribute is ProductAttachmentAttributes) {
@@ -148,9 +148,9 @@ open class ProductAttachmentViewModel : SendableViewModel,
             attachmentId: String, attachmentType: String, replyTime: String, isRead: Boolean,
             productId: Long, productName: String, productPrice: String, productUrl: String,
             productImage: String, isSender: Boolean, message: String, canShowFooter: Boolean,
-            blastId: Long, productPriceInt: Int, category: String, variants: List<AttachmentVariant>,
-            dropPercentage: String, priceBefore: String, shopId: Int, freeShipping: FreeShipping,
-            categoryId: Int, playStoreData: PlayStoreData, minOrder: Int, remainingStock: Int,
+            blastId: Long, productPriceInt: Long, category: String, variants: List<AttachmentVariant>,
+            dropPercentage: String, priceBefore: String, shopId: Long, freeShipping: FreeShipping,
+            categoryId: Long, playStoreData: PlayStoreData, minOrder: Int, remainingStock: Int,
             status: Int, wishList: Boolean, images: List<String>, source: String,
             rating: TopchatProductRating, replyId: String
     ) : super(
@@ -220,9 +220,9 @@ open class ProductAttachmentViewModel : SendableViewModel,
             attachmentId: String, attachmentType: String, replyTime: String, productId: Long,
             productName: String, productPrice: String, productUrl: String, productImage: String,
             isSender: Boolean, message: String, startTime: String, canShowFooter: Boolean,
-            blastId: Long, productPriceInt: Int, category: String, variants: List<AttachmentVariant>,
-            dropPercentage: String, priceBefore: String, shopId: Int, freeShipping: FreeShipping,
-            categoryId: Int, playStoreData: PlayStoreData, remainingStock: Int, status: Int,
+            blastId: Long, productPriceInt: Long, category: String, variants: List<AttachmentVariant>,
+            dropPercentage: String, priceBefore: String, shopId: Long, freeShipping: FreeShipping,
+            categoryId: Long, playStoreData: PlayStoreData, remainingStock: Int, status: Int,
             source: String, rating: TopchatProductRating
     ) : super(
             messageId, fromUid, from, fromRole,
@@ -283,7 +283,7 @@ open class ProductAttachmentViewModel : SendableViewModel,
     constructor(
             loginID: String, productId: Long, productName: String, productPrice: String,
             productUrl: String, productImage: String, startTime: String, canShowFooter: Boolean,
-            shopId: Int
+            shopId: Long
     ) : super(
             "", loginID, "", "",
             "", AttachmentType.Companion.TYPE_PRODUCT_ATTACHMENT, SENDING_TEXT, startTime,
@@ -397,7 +397,7 @@ open class ProductAttachmentViewModel : SendableViewModel,
     }
 
     fun isFlashSaleProduct(): Boolean {
-        return campaignId == -10000
+        return campaignId == -10000L
     }
 
     companion object {
